+++ src/Codeup/V20200414/CodeupApiResolver.php
@@ -21,6 +21,7 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method DeleteRepositoryGroup deleteRepositoryGroup(array $options = [])
  * @method DeleteRepositoryMember deleteRepositoryMember(array $options = [])
  * @method DeleteRepositoryTag deleteRepositoryTag(array $options = [])
+ * @method DeleteRepositoryWebhook deleteRepositoryWebhook(array $options = [])
  * @method GetBranchInfo getBranchInfo(array $options = [])
  * @method GetCodeupOrganization getCodeupOrganization(array $options = [])
  * @method GetFileBlobs getFileBlobs(array $options = [])
@@ -28,13 +29,17 @@ use AlibabaCloud\Client\Resolver\ApiResolver;
  * @method GetProjectMember getProjectMember(array $options = [])
  * @method GetRepositoryInfo getRepositoryInfo(array $options = [])
  * @method GetRepositoryTag getRepositoryTag(array $options = [])
+ * @method GetUserInfo getUserInfo(array $options = [])
  * @method ListGroupMember listGroupMember(array $options = [])
  * @method ListGroupRepositories listGroupRepositories(array $options = [])
  * @method ListGroups listGroups(array $options = [])
+ * @method ListOrganizations listOrganizations(array $options = [])
+ * @method ListRepositories listRepositories(array $options = [])
  * @method ListRepositoryBranches listRepositoryBranches(array $options = [])
  * @method ListRepositoryMember listRepositoryMember(array $options = [])
  * @method ListRepositoryTags listRepositoryTags(array $options = [])
  * @method ListRepositoryTree listRepositoryTree(array $options = [])
+ * @method ListRepositoryWebhook listRepositoryWebhook(array $options = [])
  * @method MergeMergeRequest mergeMergeRequest(array $options = [])
  * @method UpdateFile updateFile(array $options = [])
  * @method UpdateGroupMember updateGroupMember(array $options = [])
@@ -1063,6 +1068,49 @@ class DeleteRepositoryTag extends Roa
 
 /**
  * @method string getOrganizationId()
+ * @method string getWebhookId()
+ * @method $this withWebhookId($value)
+ * @method string getAccessToken()
+ * @method string getProjectId()
+ * @method $this withProjectId($value)
+ */
+class DeleteRepositoryWebhook extends Roa
+{
+    /** @var string */
+    public $pathPattern = '/api/v3/projects/[ProjectId]/hooks/[WebhookId]';
+
+    /** @var string */
+    public $method = 'DELETE';
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withOrganizationId($value)
+    {
+        $this->data['OrganizationId'] = $value;
+        $this->options['query']['OrganizationId'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withAccessToken($value)
+    {
+        $this->data['AccessToken'] = $value;
+        $this->options['query']['AccessToken'] = $value;
+
+        return $this;
+    }
+}
+
+/**
+ * @method string getOrganizationId()
  * @method string getSubUserId()
  * @method string getAccessToken()
  * @method string getProjectId()
@@ -1496,6 +1544,28 @@ class GetRepositoryTag extends Roa
 }
 
 /**
+ * @method string getAccessToken()
+ */
+class GetUserInfo extends Roa
+{
+    /** @var string */
+    public $pathPattern = '/api/v3/user/current';
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withAccessToken($value)
+    {
+        $this->data['AccessToken'] = $value;
+        $this->options['query']['AccessToken'] = $value;
+
+        return $this;
+    }
+}
+
+/**
  * @method string getOrganizationId()
  * @method string getSubUserId()
  * @method string getGroupId()
@@ -1784,6 +1854,176 @@ class ListGroups extends Roa
     {
         $this->data['Page'] = $value;
         $this->options['query']['Page'] = $value;
+
+        return $this;
+    }
+}
+
+/**
+ * @method string getAccessLevel()
+ * @method string getMinAccessLevel()
+ * @method string getAccessToken()
+ */
+class ListOrganizations extends Roa
+{
+    /** @var string */
+    public $pathPattern = '/api/v4/organization';
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withAccessLevel($value)
+    {
+        $this->data['AccessLevel'] = $value;
+        $this->options['query']['AccessLevel'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withMinAccessLevel($value)
+    {
+        $this->data['MinAccessLevel'] = $value;
+        $this->options['query']['MinAccessLevel'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withAccessToken($value)
+    {
+        $this->data['AccessToken'] = $value;
+        $this->options['query']['AccessToken'] = $value;
+
+        return $this;
+    }
+}
+
+/**
+ * @method string getAccessToken()
+ * @method string getArchive()
+ * @method string getSort()
+ * @method string getOrganizationId()
+ * @method string getSearch()
+ * @method string getPageSize()
+ * @method string getPage()
+ * @method string getOrder()
+ */
+class ListRepositories extends Roa
+{
+    /** @var string */
+    public $pathPattern = '/api/v3/projects/all';
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withAccessToken($value)
+    {
+        $this->data['AccessToken'] = $value;
+        $this->options['query']['AccessToken'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withArchive($value)
+    {
+        $this->data['Archive'] = $value;
+        $this->options['query']['Archive'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withSort($value)
+    {
+        $this->data['Sort'] = $value;
+        $this->options['query']['Sort'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withOrganizationId($value)
+    {
+        $this->data['OrganizationId'] = $value;
+        $this->options['query']['OrganizationId'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withSearch($value)
+    {
+        $this->data['Search'] = $value;
+        $this->options['query']['Search'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withPageSize($value)
+    {
+        $this->data['PageSize'] = $value;
+        $this->options['query']['PageSize'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withPage($value)
+    {
+        $this->data['Page'] = $value;
+        $this->options['query']['Page'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withOrder($value)
+    {
+        $this->data['Order'] = $value;
+        $this->options['query']['Order'] = $value;
 
         return $this;
     }
@@ -2174,6 +2414,72 @@ class ListRepositoryTree extends Roa
     {
         $this->data['RefName'] = $value;
         $this->options['query']['RefName'] = $value;
+
+        return $this;
+    }
+}
+
+/**
+ * @method string getOrganizationId()
+ * @method string getPageSize()
+ * @method string getAccessToken()
+ * @method string getPage()
+ * @method string getProjectId()
+ * @method $this withProjectId($value)
+ */
+class ListRepositoryWebhook extends Roa
+{
+    /** @var string */
+    public $pathPattern = '/api/v3/projects/[ProjectId]/hooks';
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withOrganizationId($value)
+    {
+        $this->data['OrganizationId'] = $value;
+        $this->options['query']['OrganizationId'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withPageSize($value)
+    {
+        $this->data['PageSize'] = $value;
+        $this->options['query']['PageSize'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withAccessToken($value)
+    {
+        $this->data['AccessToken'] = $value;
+        $this->options['query']['AccessToken'] = $value;
+
+        return $this;
+    }
+
+    /**
+     * @param string $value
+     *
+     * @return $this
+     */
+    public function withPage($value)
+    {
+        $this->data['Page'] = $value;
+        $this->options['query']['Page'] = $value;
 
         return $this;
     }